requirements.txt
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==36.0.0
cssselect==1.1.0
debugpy==1.5.1
decorator==5.1.0
Django==3.2.9
djangorestframework==3.12.4
entrypoints==0.3
fake-useragent==0.1.11
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
lxml==4.6.4
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.4
outcome==1.1.0
pandas==1.3.4
parse==1.19.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.22
pycparser==2.21
pyee==8.2.2
Pygments==2.10.0
pyOpenSSL==21.0.0
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pyzmq==22.3.0
requests==2.26.0
requests-html==0.10.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
sqlparse==0.4.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.5.2
websockets==9.1
whitenoise==5.3.0
wsproto==1.0.0
XlsxWriter==3.0.2
zipp==3.6.0
